Move ReaderResult object to ConfigReader.Result
